TS: Move Binding types to native module spec (options remain {})
