+++ js/react_native/lib/backend.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import {Backend, InferenceSession, SessionHandler, Tensor,} from 'onnxruntime-common';
+import {Backend, InferenceSession, SessionHandler, Tensor,} from '@fugood/onnxruntime-common';
 import {Platform} from 'react-native';
 
-import {binding, Binding, JSIBlob, jsiHelper} from './binding';
+import {binding, Binding, jsiHelper} from './binding';
 
 type SupportedTypedArray = Exclude<Tensor.DataType, string[]>;
 
@@ -117,7 +117,7 @@ class OnnxruntimeSessionHandler implements SessionHandler {
     const returnValue: {[name: string]: Binding.EncodedTensorType} = {};
     for (const key in feeds) {
       if (Object.hasOwnProperty.call(feeds, key)) {
-        let data: JSIBlob|string[];
+        let data: Binding.JSIBlobType|string[];
 
         if (Array.isArray(feeds[key].data)) {
           data = feeds[key].data as string[];
@@ -145,7 +145,7 @@ class OnnxruntimeSessionHandler implements SessionHandler {
         if (Array.isArray(results[key].data)) {
           tensorData = results[key].data as string[];
         } else {
-          const buffer = jsiHelper.resolveArrayBuffer(results[key].data as JSIBlob) as SupportedTypedArray;
+          const buffer = jsiHelper.resolveArrayBuffer(results[key].data as Binding.JSIBlobType) as SupportedTypedArray;
           const typedArray = tensorTypeToTypedArray(results[key].type as Tensor.Type);
           tensorData = new typedArray(buffer, buffer.byteOffset, buffer.byteLength / typedArray.BYTES_PER_ELEMENT);
         }
